requirements.txt
Django>=4.2,<5.0
python-dotenv>=1.0.0
openai>=1.0.0
markdown>=3.5
numpy>=1.24
Pillow>=10.0.0
